Applied changes requested in PR

+++ db/migrate/20240106163152_add_created_by_to_project.rb
@@ -2,6 +2,6 @@
 
 class AddCreatedByToProject < ActiveRecord::Migration[7.1]
   def change
-    add_column :projects, :created_by, :bigint
+    add_column :projects, :created_by, :integer
   end
 end
